feat: add ability to submit form

+++ dev_setup.sh
@@ -19,8 +19,10 @@ if [ -z "$NEW_UUID" ]; then
     NEW_UUID=$(uuidgen)
 fi
 
-# Create .env file with the UUID
-echo "VITE_BASE_PATH=/${NEW_UUID}" > .env
+# Create .env file with the UUID only if it doesn't exist
+if [ ! -f ".env" ]; then
+    echo "VITE_BASE_PATH=/${NEW_UUID}" > .env
+fi
 
 # Write UUID to qbid.txt
 echo "${NEW_UUID}" > qbid.txt
